Plot a curve: 32× supersampled AA, hardcoded sin placeholder
draw_curve in src/ui/plot.rs is the seam where the future Spirix-evaluated `y = f(x)` plugs in. For each interior pixel column it takes 32 centred subsamples, drops non-finite samples (NaN/±inf leave clean gaps), takes min/max of the pixel-y range, expands by ±0.5 px stroke radius, and alpha-blends each pixel row by its overlap with that span. Flat curves at fractional y get proper two-row AA; steep slopes fill many rows at full coverage with AA only at the top/bottom; samples that fall outside the view clamp to the plot interior so the curve visibly heads off the edge instead of vanishing. Hardcoded test closure plots `y = sin(x · 2π) · 0.7` until the evaluator lands. New `theme::PLOT_CURVE` is the cyan stroke colour.

Drive-by: switch mouse_x/y init from -1.0 to f32::NAN (proper sentinel — magic numbers can collide with real data). hit_test now takes f32 and rejects non-finite coords up front. Drop the dead `plot_rect.w == 0 || plot_rect.h == 0` check from update_plot_drag — it can only fire near a 2×2 window, which `resize()` already guards.

+++ src/ui/input.rs
@@ -62,7 +62,7 @@ impl PlotypusApp {
         match state {
             ElementState::Pressed => {
                 self.mouse_button_pressed = true;
-                let hit = self.hit_test(self.mouse_x as i32, self.mouse_y as i32);
+                let hit = self.hit_test(self.mouse_x, self.mouse_y);
                 match hit {
                     HIT_CLOSE_BUTTON => return ClickAction::Exit,
                     HIT_MINIMIZE_BUTTON => return ClickAction::Minimize,
@@ -125,7 +125,7 @@ impl PlotypusApp {
             self.is_dragging_resize = false;
         }
 
-        let hit = self.hit_test(x as i32, y as i32);
+        let hit = self.hit_test(x, y);
         let on_button = matches!(
             hit,
             HIT_CLOSE_BUTTON | HIT_MAXIMIZE_BUTTON | HIT_MINIMIZE_BUTTON
